2nd Part of Rails 5 Quiet Assets conversion
Disable quiet_assets gem in Gemfile. In Rails 5 the main part of quiet
asset logging is a configuration line in development.rb instead. Make
sure to use the Rails or Active...(something - prolly either Record or
Data) Logger and not the base ruby one or it won't work for that file.